+++ src/website/index.html
@@ -185,8 +185,8 @@
         >
       </footer>
       <script type="module" src="./js/index.js"></script>
-      <script src="./js/aside.js"></script>
-      <script src="./js/search-box.js"></script>
+      <script src="./js/layout/aside.js"></script>
+      <script src="./js/layout/search-box.js"></script>
     </div>
   </body>
 </html>
