+++ src/website/js/layout/search-box.js
