+++ src/website/utils/conversions.js
@@ -1,0 +1,5 @@
+const mpsToKmph = (mps) => 3.6 * mps;
+
+const kelvinToCelcius = (kelvin) => kelvin - 272.15;
+
+export default { mpsToKmph, kelvinToCelcius };
