+++ src/website/utils/dom.js
@@ -1,0 +1,5 @@
+function getEl(className) {
+  return document.querySelector("." + className);
+}
+
+export { getEl };
